rename EthUpgradeMergeList to MergeArticleList and use in beacon-change page
